refactor(android): own helper-session lifecycle and text input in their own modules
PR #1974 review, blocker 1. Both files the branch grew were already at the
extraction threshold, so the behavior landed on top of a boundary that
should have moved first.

`snapshot-helper-session.ts` (563) carried two concepts: who owns the
device's UiAutomation, and what runs over that ownership. Ownership — the
live-session registry, the enable gate, session identity, start, reuse, and
retirement — moves to `snapshot-helper-session-lifecycle.ts`. What is left
is 170 lines of snapshot capture and touch commands that acquire through it
and never reach the registry themselves; the touch path reads ownership
through `getLiveAndroidSnapshotHelperSession` instead of the map. The stale
re-export block for the retirement symbols is gone with its last importer.

`input-actions.ts` (508, pre-branch 501) sheds every text-entry path to
`text-input.ts`: provider injection, the test IME, the adb-shell writer, and
the fill orchestration this branch changed. What is left is 172 lines of
pointer, key, and gesture actions — below the pre-branch length.

Tests follow their owners. `snapshot-helper-session.test.ts` splits around
the same seam into a lifecycle file (start, reuse, identity, teardown,
quarantine) and a capture file (budgets, cancellation, fallback), and its
session fake becomes a named export of the sibling fixture module, which
drops the duplicated process double. `input-actions-fill.test.ts` and
`input-actions-test-ime.test.ts` are renamed to `text-input-*`, and the
`typeAndroid`/`fillAndroid` cases in `input-actions.test.ts` move to a new
`text-input.test.ts`.

Blocker 2. The teardown skipped `am force-stop` on an acknowledged quit plus
a clean host exit, and its own comment admitted the hole: adb without shell
protocol v2 exits 0 whenever the connection closed cleanly, including for
instrumentation that never finished, so that exit status was never proof the
device released UiAutomation. A residual-risk paragraph is not an invariant.

`adb-shell-protocol.ts` asks the transport instead. `adb features` lists
only the features both ends negotiated, so `shell_v2` there is positive
proof that a host `adb shell` child's exit status is the device command's.
The skip now requires it; an unsupported transport, a failed probe, and an
adb too old to know `features` all keep the device-side stop. The probe is
host-side and cached per device, so it costs one query per device rather
than one per teardown, and it clears with the session registry. Quarantine
keys on the same proven-release signal, so an unproven quit whose stop also
failed now reports unknown ownership instead of trusting the quit.

Red first: a transport without `shell_v2` and an unanswerable probe both
skipped the stop against the pre-fix gate, and the cache fence saw no probe
at all.

Rebased onto origin/main, which had moved the fill files.
