+++ src/platforms/android/__tests__/snapshot-helper-session.test.ts
@@ -1,119 +1,19 @@
 import assert from 'node:assert/strict';
-import { EventEmitter } from 'node:events';
-import net from 'node:net';
-import { PassThrough } from 'node:stream';
 import { afterEach, beforeEach, test } from 'vitest';
-import {
-  captureAndroidSnapshotWithHelperSession,
-  resetAndroidSnapshotHelperSessions,
-} from '../snapshot-helper-session.ts';
+import { captureAndroidSnapshotWithHelperSession } from '../snapshot-helper-session.ts';
+import { resetAndroidSnapshotHelperSessions } from '../snapshot-helper-session-lifecycle.ts';
 import { resolveAndroidSnapshotHelperSessionRequestTimeoutMs } from '../snapshot-helper-session-protocol.ts';
-import { recoverAndroidSnapshotHelperRetirement } from '../snapshot-helper-retirement.ts';
-import type { AndroidAdbExecutor, AndroidAdbProcess, AndroidAdbProvider } from '../adb-executor.ts';
+import {
+  createSessionProvider,
+  type FakeAndroidProcess,
+} from './snapshot-helper-session.fixtures.ts';
 
 beforeEach(async () => {
-  delete process.env.AGENT_DEVICE_ANDROID_SNAPSHOT_HELPER_SESSION;
   await resetAndroidSnapshotHelperSessions();
 });
 
 afterEach(async () => {
-  delete process.env.AGENT_DEVICE_ANDROID_SNAPSHOT_HELPER_SESSION;
   await resetAndroidSnapshotHelperSessions();
-});
-
-test('returns undefined when persistent sessions are disabled', async () => {
-  process.env.AGENT_DEVICE_ANDROID_SNAPSHOT_HELPER_SESSION = '0';
-  const calls: string[][] = [];
-  const provider = createSessionProvider({ calls });
-
-  const output = await captureAndroidSnapshotWithHelperSession({
-    adb: provider.exec,
-    adbProvider: provider,
-  });
-
-  assert.equal(output, undefined);
-  assert.deepEqual(calls, []);
-});
-
-test('returns undefined when the adb provider cannot spawn a helper process', async () => {
-  const calls: string[][] = [];
-  const adb: AndroidAdbExecutor = async (args) => {
-    calls.push(args);
-    return { exitCode: 0, stdout: '', stderr: '' };
-  };
-
-  const output = await captureAndroidSnapshotWithHelperSession({ adb });
-
-  assert.equal(output, undefined);
-  assert.deepEqual(calls, []);
-});
-
-test('disables repeated persistent session attempts after startup failure', async () => {
-  const calls: string[][] = [];
-  const spawnArgs: string[][] = [];
-  const provider: AndroidAdbProvider = {
-    exec: async (args) => {
-      calls.push(args);
-      return { exitCode: 0, stdout: '', stderr: '' };
-    },
-    spawn: (args) => {
-      spawnArgs.push(args);
-      const process = new FakeAndroidProcess();
-      queueMicrotask(() => process.emitExit(0, null));
-      return process;
-    },
-  };
-
-  const first = await captureAndroidSnapshotWithHelperSession({
-    adb: provider.exec,
-    adbProvider: provider,
-    deviceKey: 'android:emulator-5554',
-  });
-  const second = await captureAndroidSnapshotWithHelperSession({
-    adb: provider.exec,
-    adbProvider: provider,
-    deviceKey: 'android:emulator-5554',
-  });
-
-  assert.equal(first, undefined);
-  assert.equal(second, undefined);
-  assert.equal(spawnArgs.length, 1);
-  assert.equal(readSessionArgument(spawnArgs[0]!, 'timeoutMs'), '2000');
-  assert.equal(calls.filter((args) => args[0] === 'forward').length, 2);
-});
-
-test('starts and reuses a persistent Android snapshot helper session', async () => {
-  const calls: string[][] = [];
-  const spawnArgs: string[][] = [];
-  const provider = createSessionProvider({ calls, spawnArgs });
-
-  const first = await captureAndroidSnapshotWithHelperSession({
-    adb: provider.exec,
-    adbProvider: provider,
-    deviceKey: 'android:emulator-5554',
-    helperVersion: '0.16.2',
-    helperVersionCode: 16002,
-  });
-  const second = await captureAndroidSnapshotWithHelperSession({
-    adb: provider.exec,
-    adbProvider: provider,
-    deviceKey: 'android:emulator-5554',
-    helperVersion: '0.16.2',
-    helperVersionCode: 16002,
-  });
-
-  assert.match(first?.xml ?? '', /snapshot 1/);
-  assert.equal(first?.metadata.transport, 'persistent-session');
-  assert.equal(first?.metadata.sessionReused, false);
-  assert.equal(first?.metadata.elapsedMs, 7);
-  assert.match(second?.xml ?? '', /snapshot 2/);
-  assert.equal(second?.metadata.transport, 'persistent-session');
-  assert.equal(second?.metadata.sessionReused, true);
-  assert.equal(spawnArgs.length, 1);
-  assert.equal(
-    calls.filter((args) => args[0] === 'forward' && args[1]?.startsWith('tcp:')).length,
-    1,
-  );
 });
 
 test('allows a persistent session snapshot to use the helper command budget', async () => {
@@ -255,147 +155,6 @@ test('failed capture does not fall back when device runtime retirement is unconf
   assert.equal(processes.length, 1);
 });
 
-test('restarts the helper session when capture options change', async () => {
-  const calls: string[][] = [];
-  const spawnArgs: string[][] = [];
-  const provider = createSessionProvider({ calls, spawnArgs });
-
-  await captureAndroidSnapshotWithHelperSession({
-    adb: provider.exec,
-    adbProvider: provider,
-    deviceKey: 'android:emulator-5554',
-    waitForIdleTimeoutMs: 25,
-  });
-  const restarted = await captureAndroidSnapshotWithHelperSession({
-    adb: provider.exec,
-    adbProvider: provider,
-    deviceKey: 'android:emulator-5554',
-    waitForIdleTimeoutMs: 50,
-  });
-
-  assert.equal(restarted?.metadata.sessionReused, false);
-  assert.equal(spawnArgs.length, 2);
-  assert.equal(
-    calls.some((args) => args[0] === 'forward' && args[1] === '--remove'),
-    true,
-  );
-});
-
-test('a quit acknowledged and followed by process exit skips the force-stop round trip', async () => {
-  const calls: string[][] = [];
-  const processes: FakeAndroidProcess[] = [];
-  const provider = createSessionProvider({ calls, processes, quitExitDelayMs: 25 });
-
-  await captureAndroidSnapshotWithHelperSession({
-    adb: provider.exec,
-    adbProvider: provider,
-    deviceKey: 'android:emulator-5554',
-  });
-  await resetAndroidSnapshotHelperSessions();
-
-  assert.equal(processes.length, 1);
-  assert.equal(processes[0]?.killed, false);
-  // Acknowledged quit plus an observed exit IS the release evidence, so the extra adb round trip
-  // buys nothing.
-  assert.equal(calls.some(isHelperRuntimeForceStop), false);
-});
-
-test('a quit acknowledged by a helper the host then killed still force-stops the runtime', async () => {
-  const calls: string[][] = [];
-  const processes: FakeAndroidProcess[] = [];
-  const provider = createSessionProvider({
-    calls,
-    processes,
-    quitExit: { code: null, signal: 'SIGKILL' },
-  });
-
-  await captureAndroidSnapshotWithHelperSession({
-    adb: provider.exec,
-    adbProvider: provider,
-    deviceKey: 'android:emulator-5554',
-  });
-  await resetAndroidSnapshotHelperSessions();
-
-  // The ack alone only says the helper heard us. A host process that ended on a signal says the
-  // transport died, not that the instrumentation finished releasing UiAutomation — so the second
-  // half of the release evidence is missing and the device-side stop must still run.
-  assert.equal(calls.some(isHelperRuntimeForceStop), true);
-});
-
-test('a quit acknowledged after the host process already died still force-stops the runtime', async () => {
-  const calls: string[][] = [];
-  const processes: FakeAndroidProcess[] = [];
-  const provider = createSessionProvider({ calls, processes });
-
-  await captureAndroidSnapshotWithHelperSession({
-    adb: provider.exec,
-    adbProvider: provider,
-    deviceKey: 'android:emulator-5554',
-  });
-  // The host `am instrument` child is gone before the teardown starts, yet the device-side helper
-  // answers `quit` through the still-open forward: positive evidence that it OUTLIVED its host.
-  processes[0]?.emitExit(0, null);
-  await resetAndroidSnapshotHelperSessions();
-
-  assert.equal(calls.some(isHelperRuntimeForceStop), true);
-});
-
-test('force terminates the helper when quit is not acknowledged', async () => {
-  const calls: string[][] = [];
-  const processes: FakeAndroidProcess[] = [];
-  const provider = createSessionProvider({ calls, processes, quitResponseMode: 'malformed' });
-
-  await captureAndroidSnapshotWithHelperSession({
-    adb: provider.exec,
-    adbProvider: provider,
-    deviceKey: 'android:emulator-5554',
-  });
-  await resetAndroidSnapshotHelperSessions();
-
-  assert.equal(processes.length, 1);
-  assert.equal(processes[0]?.killed, true);
-  // Nothing proved the helper released UiAutomation, so the device-side stop must still run.
-  assert.equal(calls.some(isHelperRuntimeForceStop), true);
-});
-
-function isHelperRuntimeForceStop(args: string[]): boolean {
-  return args.join(' ') === 'shell am force-stop com.callstack.agentdevice.snapshothelper';
-}
-
-test('failed whole-module reset preserves quarantine until recovery is confirmed', async () => {
-  const options: SessionProviderOptions = {
-    calls: [],
-    quitResponseMode: 'malformed',
-    recoveryFailure: true,
-  };
-  const provider = createSessionProvider(options);
-  const deviceKey = 'android:emulator-5554';
-
-  await captureAndroidSnapshotWithHelperSession({
-    adb: provider.exec,
-    adbProvider: provider,
-    deviceKey,
-  });
-  await assert.rejects(
-    resetAndroidSnapshotHelperSessions(),
-    /Failed to retire every Android snapshot helper session/,
-  );
-  const forceStopsBeforeRecovery = options.calls.filter((args) =>
-    args.join(' ').includes('am force-stop'),
-  ).length;
-
-  options.recoveryFailure = false;
-  await recoverAndroidSnapshotHelperRetirement({
-    deviceKey,
-    adb: provider.exec,
-  });
-
-  assert.equal(
-    options.calls.filter((args) => args.join(' ').includes('am force-stop')).length,
-    forceStopsBeforeRecovery + 1,
-  );
-});
-
 test('allows device retirement beyond host-process grace before falling back', async () => {
   const calls: string[][] = [];
   const processes: FakeAndroidProcess[] = [];
@@ -432,230 +191,3 @@ test('invalidates and falls back from the helper session after a malformed respo
     true,
   );
 });
-
-type SessionProviderOptions = {
-  calls: string[][];
-  cleanupAborts?: string[][];
-  processes?: FakeAndroidProcess[];
-  quitExit?: { code: number | null; signal: NodeJS.Signals | null };
-  quitExitDelayMs?: number;
-  quitResponseMode?: 'ok' | 'malformed';
-  spawnArgs?: string[][];
-  responseMode?: 'ok' | 'malformed' | 'ui-automation-timeout';
-  responseDelayMs?: number;
-  forceStopDelayMs?: number;
-  recoveryFailure?: boolean;
-  stalledCleanup?: boolean;
-  stalledSnapshots?: number;
-};
-
-function createSessionProvider(options: SessionProviderOptions): AndroidAdbProvider {
-  let stalledSnapshots = options.stalledSnapshots ?? 0;
-  return {
-    exec: createSessionExec(options),
-    spawn: (args) => {
-      options.spawnArgs?.push(args);
-      const port = readSessionPort(args);
-      const process = new FakeAndroidProcess();
-      options.processes?.push(process);
-      let snapshotCount = 0;
-      const sockets = new Set<net.Socket>();
-      const server = net.createServer((socket) => {
-        sockets.add(socket);
-        socket.once('close', () => sockets.delete(socket));
-        socket.once('data', (chunk) => {
-          const command = chunk.toString('utf8').trim();
-          const [, requestId = ''] = command.split(/\s+/, 2);
-          if (command.startsWith('quit')) {
-            if (options.quitResponseMode === 'malformed') {
-              socket.end('not a session response');
-              return;
-            }
-            socket.end(sessionResponse({ requestId, body: '' }));
-            const quitExit = options.quitExit ?? { code: 0, signal: null };
-            server.close(() => {
-              setTimeout(
-                () => process.emitExit(quitExit.code, quitExit.signal),
-                options.quitExitDelayMs ?? 0,
-              );
-            });
-            return;
-          }
-          if (options.responseMode === 'malformed') {
-            socket.end('not a session response');
-            return;
-          }
-          if (options.responseMode === 'ui-automation-timeout') {
-            socket.end(
-              sessionResponse({
-                requestId,
-                body: '',
-                metadata: {
-                  ok: 'false',
-                  errorType: 'java.util.concurrent.TimeoutException',
-                  message: 'Timed out waiting for Android UiAutomation to connect',
-                },
-              }),
-            );
-            return;
-          }
-          snapshotCount += 1;
-          if (stalledSnapshots > 0) {
-            stalledSnapshots -= 1;
-            return;
-          }
-          const body = `<hierarchy><node text="snapshot ${snapshotCount}" /></hierarchy>`;
-          setTimeout(() => {
-            socket.end(
-              sessionResponse({
-                requestId,
-                body,
-                metadata: {
-                  waitForIdleTimeoutMs: '25',
-                  waitForIdleQuietMs: '25',
-                  timeoutMs: '5000',
-                  maxDepth: '128',
-                  maxNodes: '5000',
-                  rootPresent: 'true',
-                  captureMode: 'interactive-windows',
-                  windowCount: '1',
-                  nodeCount: '1',
-                  truncated: 'false',
-                  elapsedMs: '7',
-                },
-              }),
-            );
-          }, options.responseDelayMs ?? 0);
-        });
-      });
-      server.listen(port, '127.0.0.1', () => {
-        process.stdout.write(
-          [
-            'INSTRUMENTATION_STATUS: agentDeviceProtocol=android-snapshot-helper-v1',
-            'INSTRUMENTATION_STATUS: sessionReady=true',
-            'INSTRUMENTATION_STATUS_CODE: 2',
-            '',
-          ].join('\n'),
-        );
-      });
-      process.onKill = () => {
-        for (const socket of sockets) socket.destroy();
-        if (server.listening) {
-          server.close(() => process.emitExit(0, null));
-        } else {
-          process.emitExit(0, null);
-        }
-      };
-      return process;
-    },
-  };
-}
-
-function createSessionExec(options: SessionProviderOptions): AndroidAdbExecutor {
-  return async (args, execOptions) => {
-    options.calls.push(args);
-    const forceStopsRuntime = args.join(' ').includes('am force-stop');
-    await stallSessionCleanupIfConfigured(options, args, execOptions?.signal, forceStopsRuntime);
-    if (options.recoveryFailure && forceStopsRuntime) {
-      return { exitCode: 1, stdout: '', stderr: 'runtime still busy' };
-    }
-    await delayForceStopIfConfigured(options, execOptions?.signal, forceStopsRuntime);
-    return { exitCode: 0, stdout: '', stderr: '' };
-  };
-}
-
-async function stallSessionCleanupIfConfigured(
-  options: SessionProviderOptions,
-  args: string[],
-  signal: AbortSignal | undefined,
-  forceStopsRuntime: boolean,
-): Promise<void> {
-  const removesForward = args[0] === 'forward' && args[1] === '--remove';
-  if (!options.stalledCleanup || !signal || (!removesForward && !forceStopsRuntime)) return;
-  await new Promise<never>((_resolve, reject) => {
-    const onAbort = () => {
-      options.cleanupAborts?.push(args);
-      reject(signal.reason);
-    };
-    signal.addEventListener('abort', onAbort, { once: true });
-    if (signal.aborted) onAbort();
-  });
-}
-
-async function delayForceStopIfConfigured(
-  options: SessionProviderOptions,
-  signal: AbortSignal | undefined,
-  forceStopsRuntime: boolean,
-): Promise<void> {
-  if (!forceStopsRuntime || !options.forceStopDelayMs) return;
-  await waitForDelay(options.forceStopDelayMs, signal);
-}
-
-async function waitForDelay(delayMs: number, signal: AbortSignal | undefined): Promise<void> {
-  await new Promise<void>((resolve, reject) => {
-    const onAbort = () => {
-      clearTimeout(timer);
-      reject(signal?.reason);
-    };
-    const timer = setTimeout(() => {
-      signal?.removeEventListener('abort', onAbort);
-      resolve();
-    }, delayMs);
-    signal?.addEventListener('abort', onAbort, { once: true });
-    if (signal?.aborted) onAbort();
-  });
-}
-
-function sessionResponse(params: {
-  requestId: string;
-  body: string;
-  metadata?: Record<string, string>;
-}): string {
-  const bodyLength = Buffer.byteLength(params.body, 'utf8');
-  const headers = {
-    agentDeviceProtocol: 'android-snapshot-helper-v1',
-    helperApiVersion: '1',
-    outputFormat: 'uiautomator-xml',
-    requestId: params.requestId,
-    ok: 'true',
-    byteLength: String(bodyLength),
-    ...params.metadata,
-  };
-  return `${Object.entries(headers)
-    .map(([key, value]) => `${key}=${value}`)
-    .join('\n')}\n\n${params.body}`;
-}
-
-function readSessionPort(args: string[]): number {
-  const index = args.indexOf('sessionPort');
-  assert.notEqual(index, -1);
-  return Number(args[index + 1]);
-}
-
-function readSessionArgument(args: string[], name: string): string | undefined {
-  const index = args.indexOf(name);
-  return index < 0 ? undefined : args[index + 1];
-}
-
-class FakeAndroidProcess extends EventEmitter implements AndroidAdbProcess {
-  stdin = new PassThrough();
-  stdout = new PassThrough();
-  stderr = new PassThrough();
-  exitCode: number | null = null;
-  signalCode: NodeJS.Signals | null = null;
-  killed = false;
-  onKill: (() => void) | undefined;
-
-  kill(): boolean {
-    this.killed = true;
-    this.onKill?.();
-    return true;
-  }
-
-  emitExit(code: number | null, signal: NodeJS.Signals | null): void {
-    this.exitCode = code;
-    this.signalCode = signal;
-    this.emit('exit', code, signal);
-    this.emit('close', code, signal);
-  }
-}
